LayoutTests/accessibility/mac/aria-grid-with-strange-hierarchy-expected.txt

 	Name	Time	Popularity	Price
1	Test	Test	Test
This tests that an ARIA grid with a hierarchy that ignores render elements in between will still expose the correct elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS grid.childrenCount is 6
PASS grid.childAtIndex(0).role is 'AXRole: AXRow'
PASS grid.childAtIndex(1).role is 'AXRole: AXColumn'
PASS grid.childAtIndex(0).childrenCount is 4
PASS grid.childAtIndex(0).childAtIndex(0).role is 'AXRole: AXCell'
PASS grid.childAtIndex(0).childAtIndex(0).role is 'AXRole: AXCell'
PASS grid.childAtIndex(0).childAtIndex(0).childAtIndex(0).role is 'AXRole: AXStaticText'
PASS grid.childAtIndex(0).childAtIndex(0).childAtIndex(0).stringValue is 'AXValue: 1'
PASS successfullyParsed is true

TEST COMPLETE
